refactor: move components to their own thing
