+++ zellij-server/src/panes/grid.rs
@@ -1,9 +1,9 @@
 use super::sixel::{PixelRect, SixelGrid, SixelImageStore};
 use std::borrow::Cow;
 use std::cell::RefCell;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::HashMap;
 use std::rc::Rc;
-use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
+use unicode_width::UnicodeWidthChar;
 use zellij_utils::data::Style;
 use zellij_utils::errors::prelude::*;
 use zellij_utils::regex::Regex;
@@ -27,7 +27,7 @@ const TABSTOP_WIDTH: usize = 8; // TODO: is this always right?
 pub const MAX_TITLE_STACK_SIZE: usize = 1000;
 
 use vte::{Params, Perform};
-use zellij_utils::{consts::VERSION, shared::{version_number, ansi_len}};
+use zellij_utils::{consts::VERSION, shared::version_number};
 
 use crate::output::{CharacterChunk, OutputBuffer, SixelImageChunk};
 use crate::panes::alacritty_functions::{parse_number, xparse_color};
@@ -36,9 +36,9 @@ use crate::panes::search::SearchResult;
 use crate::panes::selection::Selection;
 use crate::panes::terminal_character::{
     AnsiCode, CharacterStyles, CharsetIndex, Cursor, CursorShape, StandardCharset,
-    TerminalCharacter, EMPTY_TERMINAL_CHARACTER, RESET_STYLES
+    TerminalCharacter, EMPTY_TERMINAL_CHARACTER
 };
-use crate::ui::boundaries::boundary_type;
+use crate::ui::components::UiComponentParser;
 
 fn get_top_non_canonical_rows(rows: &mut Vec<Row>) -> Vec<Row> {
     let mut index_of_last_non_canonical_row = None;
@@ -2131,39 +2131,6 @@ impl Grid {
     pub fn reset_cursor_position(&mut self) {
         self.cursor = Cursor::new(0, 0);
     }
-    fn parse_ui_component(&mut self, params: Vec<u8>) -> Result<()> {
-        let decoded: Vec<String> = String::from_utf8_lossy(&params).to_string().split(';').map(|c| c.to_owned()).collect();
-        let mut params_iter = decoded.iter();
-        let component_name = params_iter
-            .next()
-            .with_context(|| format!("ui component must have a name"))?;
-
-        if component_name == &"table" {
-            let columns = params_iter.next()
-                .and_then(|stringified_columns| stringified_columns.parse::<usize>().ok())
-                .with_context(|| format!("table must have columns"))?;
-            let rows = params_iter.next()
-                .and_then(|stringified_rows| stringified_rows.parse::<usize>().ok())
-                .with_context(|| format!("table must have rows"))?;
-            let stringified_params = params_iter
-                .flat_map(|stringified| {
-                    let mut utf8 = vec![];
-                    for stringified_character in stringified.split(',') {
-                        utf8.push(stringified_character.to_string().parse::<u8>().map_err(|e| format!("Failed to parse utf8: {:?}", e))?);
-                    }
-                    Ok::<String, String>(String::from_utf8_lossy(&utf8).to_string())
-                })
-                .collect::<Vec<String>>().into_iter();
-            let mut vte_parser = vte::Parser::new();
-            let encoded_table = table(columns, rows, stringified_params, Some(self.style.colors.green));
-            for &byte in &encoded_table {
-                vte_parser.advance(self, byte);
-            }
-            Ok(())
-        } else {
-            Err(anyhow!("Unknown component: {}", component_name))
-        }
-    }
 }
 
 impl Perform for Grid {
@@ -2256,7 +2223,10 @@ impl Perform for Grid {
             self.create_sixel_image();
         } else if let Some(mut ui_component_bytes) = self.ui_component_bytes.take() {
             let component_bytes = ui_component_bytes.drain(..);
-            self.parse_ui_component(component_bytes.collect()).non_fatal();
+            let style = self.style.clone();
+            UiComponentParser::new(self, style)
+                .parse(component_bytes.collect())
+                .non_fatal();
         }
         self.mark_for_rerender();
     }
@@ -3432,66 +3402,6 @@ impl Row {
         self.width = None;
         parts
     }
-}
-
-// UI COMPONENTS - TODO: move elsewhere
-fn table(columns: usize, rows: usize, contents: impl Iterator<Item=String>, title_color: Option<PaletteColor>) -> Vec<u8> {
-    let mut stringified = String::new();
-    let mut stringified_columns: BTreeMap<usize, Vec<String>> = BTreeMap::new();
-    for (i, cell) in contents.enumerate() {
-        let column_index = i % columns;
-        stringified_columns.entry(column_index).or_insert_with(Vec::new).push(cell.to_owned());
-    }
-    let mut stringified_rows: BTreeMap<usize, Vec<String>> = BTreeMap::new();
-    let mut row_width = 0;
-    for (column_index, stringified_column) in stringified_columns.values().into_iter().enumerate() {
-        let mut max_column_width = 0;
-        for cell in stringified_column {
-            let cell_width = ansi_len(cell);
-            if cell_width > max_column_width {
-                max_column_width = cell_width;
-            }
-        }
-        row_width += max_column_width + 1;
-        for (row_index, cell) in stringified_column.into_iter().enumerate() {
-            let mut padded = cell.to_owned();
-            for _ in ansi_len(cell)..max_column_width {
-                padded.push(' ');
-            }
-            stringified_rows.entry(row_index).or_insert_with(Vec::new).push(padded);
-        }
-
-    }
-    let title_styles = CharacterStyles::new()
-        .foreground(title_color.map(|t| t.into()))
-        .bold(Some(AnsiCode::On));
-    let cell_styles = CharacterStyles::new()
-        .bold(Some(AnsiCode::On));
-    for (row_index, row) in stringified_rows.values().into_iter().enumerate() {
-        let is_title_row = row_index == 0;
-        let is_last_row = row_index == rows.saturating_sub(1);
-        for cell in row {
-            if is_title_row {
-                // stringified.push_str(&format!("\u{1b}[36;1m{}\u{1b}[0m ", cell));
-                stringified.push_str(&format!("{}{}{} ", title_styles, cell, RESET_STYLES));
-            } else {
-                // stringified.push_str(&format!("\u{1b}[1m{}\u{1b}[0m ", cell));
-                stringified.push_str(&format!("{}{}{} ", cell_styles, cell, RESET_STYLES));
-            }
-        }
-        let mut title_underline = String::new();
-        for _ in 0..row_width.saturating_sub(1) { // removing 1 because the last cell doesn't have
-                                                  // padding
-            title_underline.push_str(boundary_type::HORIZONTAL);
-        }
-        if !is_last_row {
-            stringified.push_str("\n\r");
-        }
-        if is_title_row {
-            stringified.push_str(&format!("{}\n\r", title_underline));
-        }
-    }
-    stringified.as_bytes().to_vec()
 }
 
 #[cfg(test)]
