+++ zellij-server/src/ui/mod.rs
@@ -3,3 +3,4 @@ pub mod loading_indication;
 pub mod overlay;
 pub mod pane_boundaries_frame;
 pub mod pane_contents_and_ui;
+pub mod components;
